requirements.txt
pillow
torchvision
pandas